Refinements to pre-selection of candidate regions for promotion
Regions are now promoted in order of increasing amounts of live memory,
so regions with least amount of live memory get promoted first.  This
reduces the evacuation effort and expedites availability of empty
regions to improve the efficiency of future mutator allocations.  Note
that many (aged) survivor regions are poorly utilized because the mutator
prefers to allocate new objects in regions at the low end of heap memory
whereas the collector, with a much lower "allocation" rate, tends to
place survivor objects into regions at the high end of heap memory.