Tightly integrate ShenandoahThrottler within ShenandoahHeap
This is an experiment.  The observation is that there is high overhead
for all the Atomic volatile array access that was present in the original
implementation of ShenandoahThrottling.  This overhead was measured in
10-15% degradation of even P50 latencies, even when no throttling was
necessary.  Under heavy throttling conditions, this overhead manifest as
severely degraded latencies, even when no degenerated cycles were necessary.

This alternative organization should have much lower overhead:

1. Rather than two "locking" protocols at entry to each allocation
   request, we have a single protocol.  There is no atomic volatile
   gatekeeper that precedes acquisition of the heap lock.

2. Given that manipulation of the throttle queue is now under heaplock
   control, we can be more deliberate in our choice of which throttled
   threads to release when new memory becomes available, rather than
   leaving this entirely up to chance.

Let's see how this behaves.