Decommision "improved" implementation of is_good_progress
The candidate implementation used an approximation of how much memory
would be required to complete a subsequent concurrent GC as the
minimum amount of mutator free memory followwing a degenerated or
full GC.  This resulted in three Failures in the jtreg tests, with
each reporting OutOfMemory.  It may be appropriate to ask if quicker
reporting of OutOfMemory might be a more desirable behavior under
these circumstances.